Add AI Chat to all pages. Removed actual API call to reduce usage for development.
