+++ client/src/pages/Home.tsx
@@ -1,7 +1,5 @@
-import { Link, useLocation, useNavigate } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import type { Role } from "../types/Role";
-import Chatbot from "../components/Chatbot";
-import { useState } from "react";
 
 function capitalize(role: Role) {
   return role.charAt(0).toUpperCase() + role.slice(1);
@@ -24,8 +22,6 @@ export default function Home({
     setRole("guest");
     navigate("/");
   };
-  const [showChatbot, setShowChatbot] = useState(false);
-  const location = useLocation() 
 
   return (
     <section className="mb-[-0.5rem] px-1.5 sm:px-6 relative">
@@ -86,40 +82,7 @@ export default function Home({
           alt="Hero"
           className="z-[-1] hidden h-auto w-full rounded-t-[3.125rem] object-top opacity-95 sm:block"
         />
-        
-        <div className="absolute z-100 bottom-0 right-0">
-          {showChatbot && (
-            <div className="flex flex-col md:w-[600px] gap-2 px-5">
-              <Chatbot currentPage={location.pathname} role={isLoggedIn ? role : "not logged in"}></Chatbot>
-              <div className="flex justify-end">
-                <button className="h-15 w-15 rounded-full bg-accent flex justify-center items-center hover:cursor-pointer" onClick={() => setShowChatbot(false)}>
-                  <svg xmlns="http://www.w3.org/2000/svg" height="40px" viewBox="0 -960 960 960" width="40px" fill="#000000">
-                  <path d="M480-344 240-584l56-56 184 184 184-184 56 56-240 240Z"/></svg>
-                </button>
-              </div>
-              
-            </div>
-          )}
-        </div>
       </div>
-      {!showChatbot &&
-        (
-          <button className="w-30 h-30 flex items-center justify-center bg-accent rounded-l-2xl hover:cursor-pointer absolute z-100 right-0 bottom-30" onClick={() => setShowChatbot(!showChatbot)}>
-            <svg width="60" height="60" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <g clip-path="url(#clip0_4418_5763)">
-              <path d="M8.5 19H8C4 19 2 18 2 13V8C2 4 4 2 8 2H16C20 2 22 4 22 8V13C22 17 20 19 16 19H15.5C15.19 19 14.89 19.15 14.7 19.4L13.2 21.4C12.54 22.28 11.46 22.28 10.8 21.4L9.3 19.4C9.14 19.18 8.77 19 8.5 19Z" stroke="#1c0000" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-              <path opacity="0.4" d="M15.9965 11H16.0054" stroke="#1c0000" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-              <path opacity="0.4" d="M11.9955 11H12.0045" stroke="#1c0000" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-              <path opacity="0.4" d="M7.99451 11H8.00349" stroke="#1c0000" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
-              </g>
-              <defs>
-              <clipPath id="clip0_4418_5763">
-              <rect width="30" height="30" fill="none"/>
-              </clipPath>
-              </defs>
-            </svg>
-          </button>
-        )}
     </section>
   );
 }
